Move RecipesList into the API.
