+++ src/main/java/mods/eln/Eln.java
@@ -11,6 +11,7 @@ import cpw.mods.fml.common.registry.EntityRegistry;
 import cpw.mods.fml.common.registry.GameRegistry;
 import cpw.mods.fml.relauncher.Side;
 import mods.eln.api.recipe.Recipe;
+import mods.eln.api.recipe.RecipesList;
 import mods.eln.packets.*;
 import mods.eln.cable.CableRenderDescriptor;
 import mods.eln.client.ClientKeyHandler;
