+++ src/main/java/mods/eln/api/Misc.java
@@ -9,11 +9,22 @@ public class Misc {
 
     public static Object getRecipeList(String list){
         try {
-            Class<?> Eln = Class.forName("mods.eln.Eln");
-            Field Instance = Eln.getDeclaredField("instance");
-            Object instanceObject = Instance.get(null);
+            Class<?> Eln = getEln();
+            Object instanceObject = getElnInstance(Eln);
             Object recipeList = Eln.getDeclaredField(list).get(instanceObject);
             return recipeList;
+        } catch (NoSuchFieldException e) {
+            e.printStackTrace();
+        } catch (IllegalAccessException e) {
+            e.printStackTrace();
+        }
+        return null;
+    }
+
+    public static Double getElectricalFurnaceProcessEnergyNeededPerSmelt(){
+        try {
+            Class<?> ElectricalFurnaceProcess = Class.forName("mods.eln.transparentnode.electricalfurnace.ElectricalFurnaceProcess");
+            return ElectricalFurnaceProcess.getDeclaredField("energyNeededPerSmelt").getDouble(null);
         } catch (ClassNotFoundException e) {
             e.printStackTrace();
         } catch (NoSuchFieldException e) {
@@ -21,6 +32,24 @@ public class Misc {
         } catch (IllegalAccessException e) {
             e.printStackTrace();
         }
-        return null;
+        return 0D;
+    }
+
+    public static Class<?> getEln(){
+        try {
+            return Class.forName("mods.eln.Eln");
+        } catch (ClassNotFoundException e) {
+            return null;
+        }
+    }
+
+    public static Object getElnInstance(Class<?> Eln){
+        try {
+            return Eln.getDeclaredField("instance").get(null);
+        } catch (NoSuchFieldException e) {
+            return null;
+        } catch (IllegalAccessException e) {
+            return null;
+        }
     }
 }
