+++ src/main/java/mods/eln/sim/ElectricalStackMachineProcess.java
@@ -1,7 +1,7 @@
 package mods.eln.sim;
 
 import mods.eln.api.recipe.Recipe;
-import mods.eln.misc.RecipesList;
+import mods.eln.api.recipe.RecipesList;
 import mods.eln.misc.Utils;
 import mods.eln.sim.mna.component.Resistor;
 import net.minecraft.inventory.IInventory;
