+++ src/main/java/mods/eln/transparentnode/electricalmachine/ElectricalMachineDescriptor.java
@@ -4,7 +4,7 @@ import mods.eln.cable.CableRenderDescriptor;
 import mods.eln.gui.GuiLabel;
 import mods.eln.misc.*;
 import mods.eln.api.recipe.Recipe;
-import mods.eln.misc.RecipesList;
+import mods.eln.api.recipe.RecipesList;
 import mods.eln.misc.Utils;
 import mods.eln.node.transparent.TransparentNodeDescriptor;
 import mods.eln.sim.ElectricalLoad;
