+++ src/main/java/mods/eln/transparentnode/electricalmachine/MaceratorDescriptor.java
@@ -1,7 +1,10 @@
 package mods.eln.transparentnode.electricalmachine;
 
 import mods.eln.Eln;
-import mods.eln.misc.*;
+import mods.eln.misc.Direction;
+import mods.eln.misc.Obj3D;
+import mods.eln.misc.RcInterpolator;
+import mods.eln.api.recipe.RecipesList;
 import mods.eln.misc.Obj3D.Obj3DPart;
 import mods.eln.sim.ThermalLoadInitializer;
 import mods.eln.sixnode.electricalcable.ElectricalCableDescriptor;
