+++ src/main/java/mods/eln/wiki/ItemDefault.java
@@ -3,7 +3,7 @@ package mods.eln.wiki;
 import mods.eln.gui.GuiLabel;
 import mods.eln.gui.IGuiObject;
 import mods.eln.api.recipe.Recipe;
-import mods.eln.misc.RecipesList;
+import mods.eln.api.recipe.RecipesList;
 import mods.eln.misc.Utils;
 import net.minecraft.client.gui.GuiScreen;
 import net.minecraft.item.ItemStack;
